Ensure hours are extracted for the proper day
Check after extracting indices for requested hours that these hours correspond to the correct days. Otherwise, if the time requested spanned two days and the requested end time is missing but the same hour occurs on the first day, the function would plot the incorrect span of time without warning the user.